fix: fix the meaning of the state
There was confusion in the code over whether the state file contained
(a) the last day successfully written, in which case aristoteles should
start writing the subsequent day, or (b) the earliest day not yet
written, in which case aristotles should start with the day in the file.

The code now consistently uses interpretation (b).